+++ Plugins/src/main/java/tw/waterball/judgegirl/plugins/impl/cqi/CodingStyleAnalyzeReport.java
@@ -1,23 +1,35 @@
-package tw.waterball.judgegirl.entities.submission;
+/*
+ * Copyright 2020 Johnny850807 (Waterball) 潘冠辰
+ *   Licensed under the Apache License, Version 2.0 (the "License");
+ *   you may not use this file except in compliance with the License.
+ *   You may obtain a copy of the License at
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *   Unless required by applicable law or agreed to in writing, software
+ *   distributed under the License is distributed on an "AS IS" BASIS,
+ *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *   See the License for the specific language governing permissions and
+ *   limitations under the License.
+ */
+
+package tw.waterball.judgegirl.plugins.impl.cqi;
 
 import org.w3c.dom.Document;
 import org.w3c.dom.Element;
 import org.xml.sax.InputSource;
+import tw.waterball.judgegirl.entities.submission.Report;
+
 import javax.xml.parsers.DocumentBuilder;
 import javax.xml.parsers.DocumentBuilderFactory;
 import java.io.StringReader;
-import java.util.Arrays;
-import java.util.List;
+import java.util.*;
 
-import lombok.Value;
-
-@Value
-public class CodingStyleAnalyzeReport {
+public class CodingStyleAnalyzeReport extends Report {
     public String rawString;
     private Document resultXml;
     private Element xmlRootElement;
 
     public CodingStyleAnalyzeReport(String result) {
+        super("CSA-Report");
         rawString = result;
         resultXml = convertStringToXMLDocument(result);
         xmlRootElement = resultXml.getDocumentElement();
@@ -43,6 +55,16 @@ public class CodingStyleAnalyzeReport {
         } catch (Exception e) {
             throw new RuntimeException(e);
         }
+    }
+
+
+    @Override
+    public Map<String, ?> getRawData() {
+        var data = new HashMap<String, Object>();
+        data.put("score", getScore());
+        data.put("illegalName", getBadNamingStyleList());
+        data.put("globalVariable", getGlobalVariableList());
+        return data;
     }
 }
